imageAnalysis_gui: Still very much alpha.
1) Added provisional support for 32 bit multi-page tifs.

> Use importer and check multi-page, then click import images and
navigate to the image to load in the dialog box. This will load a stack
in memory called ‘importedStack’ you can make a mean projection etc in
importer for use in roiMaker. Registration won’t work yet, but will in
the next update. Overall support will be seamless next time.

2) Changed the behavior of roiMaker’s ‘Global XCor’ button. This is the
button used to create a global cross correlation that gets used for the
local stuff. What it will do now is by default look to see if a stack
named ‘importedStack’ is in the workspace. If ‘importedStack’ exists it
will use it, if not it will throw up a dialog and ask you for a folder
full of individual tifs.

> This is a temporary kludge to enable basic function right now. I will
enable import of a multi-page tif or individual tifs so all is seamless
soon. Also, I will allow a user to select the stack to use from the
workspace variable dialog box.