Version 0.5
Importer is improved:
1) Control over parallelization on import or registration.
2) You can now easily apply reg transforms to different stacks.
3) Removed the concept of secondary registration stack, instead apply
transforms.
4) You can quickly plot a projection or movie to inspect things.
5) I think I have resolved all uint16/double conversion stuff. I do not
think uint32 is as debugged.

roiMaker is improved:
1) XCorr now works on the stack you select in the workspace var box.
2) You can no longer compute xcorrs from disk, instead import a small
stack in importer.
3) There is a lot more to do here. Next step is adding in the concept
of groups. So you can have a soma, dendrite and axon in the same group.
You will also be able to have different channels.
4) I will add an optional PCA step to help refine ROIs.

roiInspector is improved:
1) There is now a workspace var box that lets you select where to
extract from.
Caveats:
If you extract from multiple stacks, it currently just writes over. I
left this the way it was because I will have to modify with the whole
group concept.
**** Extract Button works as advertised. The other extract type buttons
are broken.