2.8 - Event Inspector; Minor UI Bug Fixes
added eventInspector:

eventInspector is a program that will allow you to load a 2-D matrix of
values and inspect each element (cell) as a whole trace (trial), or as
individual portion (event/trial). Trials or events are passed as a
vector of times in seconds. Timing data is hand coded in the gui, or
you can load from a struct called ‘timingParams.’  After you load the
initial data it remembers what you were working with (e.g. somaticF
only has to be specified in the workspace browser to let the ‘load
data’ button know that is what you want).

Known Issues:
1) Timing of trial is fixed at 8 seconds. This is temporary and will be
fixed next update.

2) Frame interval can be specified in the GUI. However, because I am
working on some precision issues, I have temporarily hardcoded this as
my typical frame interval. This will be fixed next update too.

3) I have not put in the ability to save out the timing params set in
the gui, but I will next update.

4) PSTHs will be in next update.