v1.25 -- Significant HDF Improvments
a) HDF5 across platforms can lead to stacks rendered with either the z dimension (frame number) as the first or last entry of the data dimensions (typically). In python's h5py, if you follow typical practices, you will end up with z in the first dimension, but MATLAB will lead to the third. I added a "zDim=1" toggle so you can tell MATLAB to flip the z-dimension. You will see the need for this if the end image displayed is not what you expect.

b) constraining the import indices (end image; first image) now work with hdf5.