v1.5 -- Many Such Improvements
Improvments? Many such cases.
This is a substantial update, most of the changes are in roiMaker and importer.

I think I have killed 99% of the interface shifting bugs in roiMaker.

I now predict what slider values you are likley to want across image types.

roiOverlay works with movies, it does not overlay all of your rois, but instead the last active mask, which is cMask in your workspace. You can set this by hand to anything.

Filters work during playback.

After a movie plays there should be no frameNumber invalid bugs.

Removed superfulous buttons (not all).

Better memory management in importer and roiMaker.

stack averaging in importer defaults to shave your stack at the end to give you the division you seek. So if you have a stack of 999 and want to average by 3, well that was easy before and continues to give you 333. But, if you want to average by 4 then you would need a stack of 996, I shave and then average to 249.

I think I have fixed any place that converts a uint16 stack, unecessarily to a double.

nnmf in roiMaker now outputs nm1 and nm1Nrm stacks. The nm1Nrm is nm1, but normalized to the in frame max, instead of all predicted variance.

a bunch of other stuff I can't remember.

extractor will get the same overhaul in the coming days. a big speed boost is in the offing.

-chris