1.65 -- Most roiMaker ui bugs killed.
roiMaker now rolls over the frame count when you cut through a stack, or play one as a movie.

The roi overlay filter works while playing a movie. It defaults to whatever is cMask.

There is an xy crop bound box in roiMaker now and when playing a movie, it will show the cropped region, and any other selected filter on top of it.