v1.7b -- Changes to Importer
Added Auto Import, which does the following:

a) Imports the images as you specify
b) Computes the mean of all those images
c) Makes a template from N number of images that are in the lowest decile of the images you imported. N is the number of images next to "auto template" in stack maniupulations.
d) Registers a stack called "importedImages" which is the default name.